+++ .gitignore
@@ -1,5 +1,6 @@
 node_modules/
-
+assets/
+dist/
 output/
 *.js
