+++ .npmignore
@@ -1,0 +1,3 @@
+assets/
+output/
+node-modules/
