--- src/types/ModelTypes.ts
@@ -1,5 +1,0 @@
-export enum ModelType {
-    OBJECT = 1,
-    SKIN = 2,
-    CAR = 3
-  }
