+++ src/types/model-types.ts
@@ -1,0 +1,5 @@
+export enum ModelType {
+    OBJECT = "object",
+    SKIN = "skin",
+    CAR = "car"
+  }
