+++ src/utils/geometry-utils.ts
