+++ src/utils/matrix-utils.ts
